refactor(resources): update dto
