+++ src/share-resources/dto/shared-resource.dto.ts
@@ -1,0 +1,20 @@
+import { ResourceType } from 'omniboxd/resources/resources.entity';
+import { Resource } from 'omniboxd/resources/resources.entity';
+
+export class SharedResourceDto {
+  id: string;
+  name: string;
+  resource_type: ResourceType;
+  content: string;
+  attrs: Record<string, any>;
+
+  static fromEntity(resource: Resource) {
+    const dto = new SharedResourceDto();
+    dto.id = resource.id;
+    dto.name = resource.name;
+    dto.resource_type = resource.resourceType;
+    dto.content = resource.content;
+    dto.attrs = resource.attrs;
+    return dto;
+  }
+}
